+++ src/searchRecap/SearchRecap.unit.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+
+import SearchRecap from './SearchRecap'
+
+const defaultProps = {
+  from: 'Middlesbrough',
+  to: 'Stoke-on-Trent, city center',
+  info: 'Today 2:30pm, 2 passengers',
+  large: false,
+}
+
+describe('searchRecap', () => {
+  it('Should display searchRecap component', () => {
+    const component = renderer.create(<SearchRecap {...defaultProps} />).toJSON()
+    expect(component).toMatchSnapshot()
+  })
+})